Title: 특정 문자로 시작하는 문자열, Time: 101ms, Memory: 9MB, Status: Passed - CodeTree

+++ 231224/특정 문자로 시작하는 문자열/strings-that-start-with-a-specific-character.java
@@ -31,7 +31,7 @@ public class Main {
             }
         }
 
-        double avg = sum/cnt;
+        double avg = (double)sum/cnt;
         System.out.printf("%d %.2f", cnt, avg);
     }
 }
